task 4.5: updating rtime, stime , retime for each process in function update_times(void)

+++ proc.h
@@ -11,6 +11,7 @@ struct cpu
   struct proc *proc;         // The process running on this cpu or null
 };
 
+
 extern struct cpu cpus[NCPU];
 extern int ncpu;
 
@@ -64,9 +65,9 @@ struct proc
   long long accumulator; // Added for task 4.2
   int ps_priority;
   double cfs_priority;      //decay factor
-  long long rtime;             //time process was in running state.
-  long long stime;             //time process was in sleeping state.
-  long long retime;            //time process was in ready
+  int rtime;             //time process was in running state.
+  int stime;             //time process was in sleeping state.
+  int retime;            //time process was in ready
 
 };
